Added Video Option for Chipset graphics that cannot create big textures.
- no low-fi version of tables required anymore.

As you know i have a bad graphic card that cannot create textures > 1024. Textures are now resized when a table is loaded (and that feature is turned on in Video Options). This feature only works for images != bmp (since bmp is stored lzw compressed in VPT-files).

I'll have a look at this later.

This was a hard one, since i first tried to resize Pictures when rendering takes place -> this did not work because textures were already scrambled at loading time.

git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@181 14438d79-0ffd-4333-8ae7-ea2bdd901387